refactor[code]: Done refactoring of the card , sidebar , navbar , hero and main content.
[Added the fixed layout for the sidebar and made it fixed.
Changed the background of the hero section and done major refactoring.
Fixed up the navbar at the top.
Change the UI of the main sections of equipments , body parts , exercises , target muscles and routines page.
Added a different card for all the content pages.
Done refactoring of the content pages by making a single card component.]

Signed-off-by: Akshat Jaiswal <akshatjaiswal.official@gmail.com>